OBPIH-6984 Improve validations on existing inventories on count step  (#5066)
